[MOVEONLY] Move CFeeRate out of the consensus module
>>> backports bitcoin/bitcoin@381a46e38fa9262c868e5fee9ed0d1f71af4059e
