+++ src/test/policyestimator_tests.cpp
@@ -2,6 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "policy/feerate.h"
 #include "policy/fees.h"
 #include "txmempool.h"
 #include "uint256.h"
