+++ src/txmempool.h
@@ -13,6 +13,7 @@
 
 #include "amount.h"
 #include "coins.h"
+#include "policy/feerate.h"
 #include "primitives/transaction.h"
 #include "sync.h"
 #include "random.h"
